+++ Services/UserService/IUserService.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace insulin_backend.Services.UserService
+{
+    public interface IUserService
+    {
+        Object GetAllUsersTutorials(int userId);
+    }
+}
